megawifi: move sources and includes to src/ext/mw and src/inc/mw
