+++ sample/megawifi/src/main.c
@@ -18,7 +18,7 @@
 #include <types.h>
 #include <genesis.h>
 #include <tsk.h>
-#include <megawifi.h>
+#include <mw/megawifi.h>
 
 #if (ENABLE_MULTITASK == 0) || (ENABLE_MEGAWIFI == 0)
 #error "Set ENABLE_MULTITASK and ENABLE_MEGAWIFI to 1 in config.h and rebuild the library"
